Started work on LESS refactoring
