--- src/components/Checkmark/checkmark.css
@@ -1,123 +1,0 @@
-.checkmark-display {
-  position: absolute; 
-  /* Right aligned solid check */
-  right: 0;
-  padding-bottom: 3px; 
-  padding-right: 4px; 
-  /* Center-aligned check above submit */
-  /* margin-top: 85px;
-  left: 45%; */
-}
-
-.checkmark-hide {
-  display: none;
-}
-
-/* Right aligned solid check */
-
-.checkmark {
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  display: block;
-  stroke-width: 2;
-  stroke: #fff;
-  stroke-miterlimit: 10;
-  box-shadow: inset 0px 0px 0px #7ac142;
-  animation: fill .4s ease-in-out .4s forwards, scale .3s ease-in-out .9s both;
-}
-
-.checkmark__circle {
-  stroke-dasharray: 166;
-  stroke-dashoffset: 166;
-  stroke-width: 2;
-  stroke-miterlimit: 10;
-  stroke: #7ac142;
-  fill: none;
-  animation: stroke 0.6s cubic-bezier(0.65, 0, 0.45, 1) forwards;
-}
-
-.checkmark__check {
-  transform-origin: 50% 50%;
-  stroke-dasharray: 48;
-  stroke-dashoffset: 48;
-  animation: stroke 0.3s cubic-bezier(0.65, 0, 0.45, 1) 0.8s forwards;
-}
-
-@keyframes stroke {
-  100% {
-    stroke-dashoffset: 0;
-  }
-}
-@keyframes scale {
-  0%, 100% {
-    transform: none;
-  }
-  50% {
-    transform: scale3d(1.1, 1.1, 1);
-  }
-}
-@keyframes fill {
-  100% {
-    box-shadow: inset 0px 0px 0px 30px #7ac142;
-  }
-}
-
-/* Center-aligned check above submit */
-
-/* svg {
-  width: 50px;
-  display: block;
-  margin: 40px auto 0;
-  animation-duration: 2s;
-}
-.path {
-  stroke-dasharray: 1000;
-  stroke-dashoffset: 0;
-}
-.path.circle {
-  -webkit-animation: dash 1.4s ease-in-out;
-  animation: dash 1.4s ease-in-out;
-}
-.path.line {
-  stroke-dashoffset: 1000;
-  -webkit-animation: dash 1.4s 0.5s ease-in-out forwards;
-  animation: dash 1.4s 0.5s ease-in-out forwards;
-}
-.path.check {
-  stroke-dashoffset: -100;
-  -webkit-animation: dash-check 1.4s 0.5s ease-in-out forwards;
-  animation: dash-check 1.4s 0.5s ease-in-out forwards;
-}
-@-webkit-keyframes dash {
-  0% {
-    stroke-dashoffset: 1000;
-  }
-  100% {
-    stroke-dashoffset: 0;
-  }
-}
-@keyframes dash {
-  0% {
-    stroke-dashoffset: 1000;
-  }
-  100% {
-    stroke-dashoffset: 0;
-  }
-}
-@-webkit-keyframes dash-check {
-  0% {
-    stroke-dashoffset: -100;
-  }
-  100% {
-    stroke-dashoffset: 900;
-  }
-}
-@keyframes dash-check {
-  0% {
-    stroke-dashoffset: -100;
-  }
-  100% {
-    stroke-dashoffset: 900;
-  }
-} */
