+++ src/components/Verb/Settings/Settings.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import VerbTenses from './VerbTenses';
 import NumberPerson from './VerbNumberPerson';
-import '../../../app.css';
+import '../../../css/app.css';
 
 class Settings extends Component {
   constructor(props) {
