+++ src/css/modal.css
@@ -2,13 +2,12 @@
   display: flex;
   justify-content: center;
 }
-
 .modal-wrapper {
   background: white;
   border: 1px solid #d0cccc;
-  -webkit-box-shadow: 0 3px 7px rgba(208,204,204,0.29);
-  -moz-box-shadow: 0 3px 7px rgba(208,204,204,0.29);
-  box-shadow: 0 3px 7px rgba(208,204,204,0.29);
+  -webkit-box-shadow: 0 3px 7px rgba(208, 204, 204, 0.29);
+  -moz-box-shadow: 0 3px 7px rgba(208, 204, 204, 0.29);
+  box-shadow: 0 3px 7px rgba(208, 204, 204, 0.29);
   transition: all 1.3s;
   max-width: 380px;
   position: absolute;
@@ -18,9 +17,7 @@
   z-index: 5;
   margin: 0 auto;
 }
-
 .modal-header {
-  height: 80px;
   line-height: 40px;
   padding: 0 20px 5px 20px;
   text-align: center;
@@ -28,35 +25,29 @@
   height: 80px;
   border-bottom: #263238 solid 2px;
 }
-
 .modal-header h3 {
   color: black;
   float: left;
   margin: 0;
   padding: 0;
 }
-
 .logo {
   font-family: 'Saira', cursive;
   color: blue;
 }
-
 .modal-body {
   padding: 10px 15px;
   text-align: center;
 }
-
 .modal-footer {
   background: #263238;
   height: 28px;
   padding: 15px;
 }
-
 .btn-cancel {
   border: none;
   outline: none;
   background-color: red;
-  float: left;
   font-size: 0.8rem;
   font-family: 'Open Sans';
   border-radius: 5px;
@@ -66,16 +57,13 @@
   float: left;
   cursor: pointer;
 }
-
 .created-by {
-  display: flex
+  display: flex;
 }
-
 .created-by img {
   width: 100px;
   height: 100px;
 }
-
 .created-by a {
   color: blue;
 }
