+++ src/less/app.less
@@ -1,47 +1,37 @@
 @import url('https://fonts.googleapis.com/css?family=Open+Sans:400,700,800');
 @import url('https://fonts.googleapis.com/css?family=Saira');
 
-/* Global styles */
+@import 'variables';
+
+
 
 body {
   font-family: 'Open Sans';
   background-color: #f6f9fc;
+
+  @media @mobile630 {
+      width: 95vw;
+   }
 }
 
 .app-wrapper {
   margin-top: 10px;
   padding: 10px 0;
   background-color: #f6f9fc;
-}
+ 
+ @media @desktop630 {
+      width: 580px;
+      margin: 0 auto;
+    }
 
-/* Desktop styles */
+ @media @mobile630 {
+     width: 380px;
+     margin: 0 auto;
+    }
 
-@media only screen and (min-width: 630px) {
-  .app-wrapper {
-    width: 580px;
-    margin: 0 auto;
-  }
-}
-
-/* Mobile styles */
-
-@media only screen and (max-width: 630px) {
-  body {
+ @media @mobile450 {
     width: 95vw;
-  }
-}
-
-@media only screen and (max-width: 630px) {
-  .app-wrapper {
-    width: 380px;
-    margin: 0 auto;
-  }
-}
-
-@media only screen and (max-width: 450px) {
-  .app-wrapper {
-    width: 95vw;
-  }
+    }
 }
 
 /* Header styles */
@@ -51,28 +41,27 @@ body {
   background: white;
   height: 90px;
   padding-top: 5px;
-}
 
-@media only screen and (max-width: 390px) {
-  .header {
+  @media only screen and (max-width: 390px) {
     height: 130px;
+    }
+
+  h2 {
+    font-size: 20px;
+    margin: 10px 0;
+    font-weight: 400;
   }
-}
 
-.header h2 {
-  font-size: 20px;
-  margin: 10px 0;
-  font-weight: 400;
-}
+  span {
+    font-family: 'Saira', cursive;
+    color: @darkblue;
+    font-weight: 700;
+  }
 
-.header span {
-  font-family: 'Saira', cursive;
-  color: #003893;
-  font-weight: 700;
-}
+  div {
+    font-size: 16px;
+  }
 
-.header div {
-  font-size: 16px;
 }
 
 /* Verb streak styles */
@@ -113,10 +102,9 @@ body {
   justify-content: flex-start;
   text-align: center;
   margin-bottom: 7.5px;
-}
-
-.person-flex div {
-  font-weight: 800;
+  div {
+    font-weight: 800;
+  }
 }
 
 .input-section {
@@ -127,10 +115,10 @@ body {
   align-items: flex-end;
   margin: 7% 0% 2% 0%;
   position: relative;
-}
 
-.input-section input {
-  padding-left: 10px;
+  input {
+    padding-left: 10px;
+  }
 }
 
 .input {
@@ -151,7 +139,7 @@ body {
 /* Accent button styles */
 
 .accent-buttons button {
-  background-color: #003893;
+  background-color: @darkblue;
   color: white;
   border-radius: 4px;
   margin: 0 2px 6px 2px;
@@ -164,10 +152,9 @@ body {
   height: 85px;
   position: relative;
   padding: 10px 0;
-}
-
-.helper-text div {
-  text-align: center;
+  div {
+    text-align: center;
+  }
 }
 
 /* Submit button styles */
@@ -178,7 +165,7 @@ body {
   border-radius: 5px;
   padding: 10px 15px;
   color: white;
-  background-color: #003893;
+  background-color: @darkblue;
   width: 70%;
   position: absolute;
   z-index: 0;
@@ -196,11 +183,10 @@ body {
   font-size: 24px;
   font-weight: 700;
   text-transform: uppercase;
-}
-
-.twenty-four span {
-  font-weight: 400;
-  text-transform: none;
+  span {
+    font-weight: 400;
+    text-transform: none;
+  }
 }
 
 /* Settings styles */
@@ -247,10 +233,9 @@ body {
 
 .hover-text { 
   cursor: pointer;
-}
-
-.hover-text:hover {
-  text-decoration: underline;
+  &:hover {
+    text-decoration: underline;
+  }
 }
 
 .made-with-love {
