+++ src/less/variables.less
@@ -1,0 +1,6 @@
+
+@desktop630: ~"only screen and (min-width: 630px)";
+@mobile450: ~"only screen and (max-width: 450px)";
+@mobile630: ~"only screen and (max-width: 630px)";
+
+@darkblue: #003893;
